requirements.txt
python-docx
mammoth
odfpy
jinja2
beautifulsoup4